[NEW] Criando o carrinho de compras

+++ js/index.js
@@ -1,10 +1,16 @@
 //Salvar produto no LocalStorage
 var id = 1
+
+$(function(){
+    listProducts()
+})
+
 var save_product = document.getElementById("save_product")
 save_product.addEventListener("click", function(){
     var product_name = document.getElementById("name").value
     var product_brand = document.getElementById("brand").value
     var product_price = document.getElementById("price").value
+    var product_bar_code = document.getElementById("bar_code").value
 
     //verificar se localstorage existe
     if(localStorage.getItem("products") != null)
@@ -12,6 +18,7 @@ save_product.addEventListener("click", function(){
         var deserialize_products = JSON.parse(localStorage.getItem("products"))
         var new_product = {
             "id":id,
+            "bar_code":product_bar_code,
             "product_name": product_name,
             "product_brand":product_brand,
             "product_price":product_price
@@ -24,6 +31,7 @@ save_product.addEventListener("click", function(){
         var initial_product = [
             {
                 "id":id,
+                "bar_code":product_bar_code,
                 "product_name": product_name,
                 "product_brand":product_brand,
                 "product_price":product_price
@@ -37,18 +45,63 @@ save_product.addEventListener("click", function(){
 
 })
 
-$(function(){
-    listProducts()
-})
 
 function deleteProduct(id_produto)
 {
-    alert(id_produto)
+    var deserialized_local_storage = JSON.parse(localStorage.getItem("products"))
+    var array = []
+    deserialized_local_storage.forEach(obj => {
+        if (obj.id != id_produto){
+            array.push(obj)
+        }
+    });
+
+    deserialized_local_storage = array
+    localStorage.setItem("products", JSON.stringify(deserialized_local_storage))
+    listProducts()
 }
 
 function addToChart(produto)
 {
-    console.log(produto)
+    if(localStorage.getItem("cart") != null){
+        //verificar se o produto já existe no carrinho
+        var deserialized_cart = JSON.parse(localStorage.getItem("cart"))
+        var situation = true
+
+        deserialized_cart.forEach(obj => {
+            if(obj.produto.bar_code == produto.bar_code)
+            {
+                situation = false
+            }
+        })
+
+        if(situation)
+        {
+            deserialized_cart.push(produto)
+        }
+        else
+        {
+            deserialized_cart.forEach(obj => {
+                if(obj.produto.bar_code == produto.bar_code)
+                {
+                    obj.quantidade = obj.quantidade + 1
+                }
+            })
+        }
+
+        localStorage.setItem("cart", JSON.stringify(deserialized_cart))
+
+    }else{
+        var first_product = [
+            {
+                produto,
+                "quantidade":1
+            }
+        ]
+
+        localStorage.setItem("cart", JSON.stringify(first_product))
+    }
+    alert(`${produto.product_name} Adicionado com succeso ao carrinho`)
 }
 
 //Listar produtos na tabela
@@ -63,6 +116,9 @@ function listProducts(){
         //create <th></th>
         var td_id = document.createElement("td")
         td_id.appendChild(document.createTextNode(obj.id))
+
+        var td_bar_code = document.createElement("td")
+        td_bar_code.appendChild(document.createTextNode(obj.bar_code))
 
         var td_name = document.createElement("td")
         td_name.appendChild(document.createTextNode(obj.product_name))
@@ -79,7 +135,10 @@ function listProducts(){
         button_delete.classList.add("btn-danger")
 
         button_delete.addEventListener("click", function(){
-            deleteProduct(obj.id)
+            if(confirm(`Deseja mesmo Remover ${obj.product_name} da Lista de Produtos ?`))
+            {
+                deleteProduct(obj.id)
+            }
         })
 
         button_delete.appendChild(document.createTextNode("Delete"))
@@ -89,12 +148,15 @@ function listProducts(){
         button_add_chart.classList.add("btn-primary")
         button_add_chart.appendChild(document.createTextNode("+ Comprar"))
         button_add_chart.addEventListener("click", function(){
-            addToChart(obj)
+            if(confirm(`Deseja mesmo adicionar ${obj.product_name} ao Carrinho?`)){
+                addToChart(obj)
+            }
         })
         td_action.appendChild(button_add_chart)
         td_action.appendChild(button_delete)
 
         tr.appendChild(td_id)
+        tr.appendChild(td_bar_code)
         tr.appendChild(td_name)
         tr.appendChild(td_brand)
         tr.appendChild(td_price)
